+++ packages/expect/package.json
@@ -1,12 +1,12 @@
 {
-  "name": "truffle-expect",
+  "name": "@truffle/expect",
   "version": "0.0.9",
   "description": "Simple module for ensuring specific options are passed to a function",
   "main": "index.js",
   "scripts": {
     "test": "mocha"
   },
-  "repository": "https://github.com/trufflesuite/truffle/tree/master/packages/truffle-expect",
+  "repository": "https://github.com/trufflesuite/truffle/tree/master/packages/expect",
   "keywords": [
     "truffle",
     "ethereum",
@@ -17,7 +17,7 @@
   "bugs": {
     "url": "https://github.com/trufflesuite/truffle/issues"
   },
-  "homepage": "https://github.com/trufflesuite/truffle/blob/master/packages/truffle-expect#readme",
+  "homepage": "https://github.com/trufflesuite/truffle/blob/master/packages/expect#readme",
   "publishConfig": {
     "access": "public"
   },
