+++ packages/truffle-compile/new/index.js
@@ -1,7 +1,7 @@
 const debug = require("debug")("compile:new"); // eslint-disable-line no-unused-vars
 const path = require("path");
 const { promisify } = require("util");
-const expect = require("truffle-expect");
+const expect = require("@truffle/expect");
 const findContracts = require("truffle-contract-sources");
 const Config = require("truffle-config");
 const Profiler = require("../profiler");
