+++ packages/truffle-core/lib/console.js
@@ -4,7 +4,7 @@ const provision = require("truffle-provisioner");
 const contract = require("truffle-contract");
 const { Web3Shim } = require("truffle-interface-adapter");
 const vm = require("vm");
-const expect = require("truffle-expect");
+const expect = require("@truffle/expect");
 const TruffleError = require("truffle-error");
 const fse = require("fs-extra");
 const path = require("path");
