+++ packages/truffle-core/lib/test.js
@@ -9,7 +9,7 @@ const TestRunner = require("./testing/testrunner");
 const TestResolver = require("./testing/testresolver");
 const TestSource = require("./testing/testsource");
 const SolidityTest = require("./testing/soliditytest");
-const expect = require("truffle-expect");
+const expect = require("@truffle/expect");
 const Migrate = require("truffle-migrate");
 const Profiler = require("truffle-compile/profiler");
 const originalrequire = require("original-require");
