+++ packages/truffle-core/lib/testing/testrunner.js
@@ -3,7 +3,7 @@ var Config = require("truffle-config");
 var Migrate = require("truffle-migrate");
 var TestResolver = require("./testresolver");
 var TestSource = require("./testsource");
-var expect = require("truffle-expect");
+var expect = require("@truffle/expect");
 var contract = require("truffle-contract");
 var abi = require("web3-eth-abi");
 var path = require("path");
