+++ packages/truffle-deployer/package.json
@@ -2,7 +2,7 @@
   "name": "truffle-deployer",
   "description": "Light deployment module for easily deploying Ethereum contracts",
   "license": "MIT",
-  "author": "Tim Coulter <tim.coulter@consensys.net>",
+  "author": "Tim Coulter <tim@trufflesuite.com>",
   "homepage": "https://github.com/trufflesuite/truffle-deployer#readme",
   "repository": "https://github.com/trufflesuite/truffle/tree/master/packages/truffle-deployer",
   "bugs": {
@@ -16,7 +16,7 @@
   "dependencies": {
     "emittery": "^0.4.0",
     "truffle-contract": "^4.0.30",
-    "truffle-expect": "^0.0.9"
+    "@truffle/expect": "^0.0.9"
   },
   "devDependencies": {
     "ganache-core": "2.7.0",
