+++ packages/truffle-environment/environment.js
@@ -1,7 +1,7 @@
 const Web3 = require("web3");
 const Web3Shim = require("truffle-interface-adapter").Web3Shim;
 const TruffleError = require("truffle-error");
-const expect = require("truffle-expect");
+const expect = require("@truffle/expect");
 const Resolver = require("truffle-resolver");
 const Artifactor = require("truffle-artifactor");
 const Ganache = require("ganache-core/public-exports");
