+++ packages/truffle-environment/package.json
@@ -16,7 +16,7 @@
     "node-ipc": "^9.1.1",
     "truffle-artifactor": "^4.0.29",
     "truffle-error": "^0.0.5",
-    "truffle-expect": "^0.0.9",
+    "@truffle/expect": "^0.0.9",
     "truffle-interface-adapter": "^0.2.4",
     "truffle-resolver": "^5.0.15",
     "web3": "1.2.1"
