+++ packages/truffle-external-compile/index.js
@@ -6,7 +6,7 @@ const resolve = require("path").resolve;
 const { callbackify, promisify } = require("util");
 const glob = promisify(require("glob"));
 const fs = require("fs");
-const expect = require("truffle-expect");
+const expect = require("@truffle/expect");
 const Schema = require("truffle-contract-schema");
 const web3 = {};
 web3.utils = require("web3-utils");
