+++ packages/truffle-migrate/index.js
@@ -2,7 +2,7 @@ const fs = require("fs");
 const dir = require("node-dir");
 const path = require("path");
 const async = require("async");
-const expect = require("truffle-expect");
+const expect = require("@truffle/expect");
 const Config = require("truffle-config");
 const Reporter = require("truffle-reporters").migrationsV5;
 const Migration = require("./migration.js");
