+++ packages/truffle-require/package.json
@@ -16,7 +16,7 @@
   "dependencies": {
     "original-require": "1.0.1",
     "truffle-config": "^1.1.19",
-    "truffle-expect": "^0.0.9",
+    "@truffle/expect": "^0.0.9",
     "truffle-interface-adapter": "^0.2.4",
     "web3": "1.2.1"
   },
