+++ packages/truffle-workflow-compile/utils.js
@@ -1,5 +1,5 @@
 const Config = require("truffle-config");
-const expect = require("truffle-expect");
+const expect = require("@truffle/expect");
 const Resolver = require("truffle-resolver");
 const Artifactor = require("truffle-artifactor");
